lichess-opening-builder/engine_based/requirements.txt
# Base requirements (from parent directory)
requests>=2.28.0
python-dotenv>=1.0.0
chess>=1.10.0

# Maia2 and its requirements
maia2
torch>=2.4.0
numpy>=2.1.3
einops>=0.8.0
gdown>=5.2.0
pandas>=2.2.3
pyzstd>=0.15.9
tqdm>=4.65.0
PyYAML>=6.0.0

# Note: Stockfish must be installed separately via system package manager
# Fedora: sudo dnf install stockfish
# Ubuntu: sudo apt install stockfish
# macOS: brew install stockfish
